move ForEachVariant to shapely crate
