--- lib/rust/prelude/macros/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "enso-prelude-macros"
-version = "0.1.0"
-edition = "2021"
-authors = ["Enso Team <contact@enso.org>"]
-
-[lib]
-proc-macro = true
-
-[dependencies]
-Inflector = "0.11.4"
-proc-macro2 = "1.0"
-quote = "1.0"
-syn = { version = "1.0", features = [
-  "full",
-  "extra-traits",
-  "printing",
-  "parsing"
-] }
