+++ lib/rust/prelude/src/macros.rs
@@ -1,6 +1,11 @@
 //! This module defines set of common macros which are useful across different projects.
 
-pub use enso_prelude_macros::*;
+
+// ==============
+// === Export ===
+// ==============
+
+pub use enso_shapely::ForEachVariant;
