+++ lib/rust/shapely/macros/src/lib.rs
@@ -23,6 +23,7 @@ extern crate proc_macro;
 
 mod derive_clone_ref;
 mod derive_entry_point;
+mod derive_for_each_variant;
 mod derive_iterator;
 mod derive_no_clone;
 mod overlappable;
@@ -112,6 +113,38 @@ pub fn derive_clone_ref(input: proc_macro::TokenStream) -> proc_macro::TokenStre
 #[proc_macro_derive(NoCloneBecauseOfCustomDrop)]
 pub fn derive_no_clone(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     derive_no_clone::derive(input)
+}
+
+/// Implements the `ForEachVariant` derive macro which creates a helper for iterating over each
+/// variant of an enum at compile time. The derive panics if used on non-enum types.
+///
+/// The derive creates a macro (hereafter called loop-macro) named `for_each_NAME_variant` where
+/// `NAME` is replaced with the name of the enum converted to snake case. The loop-macro takes a
+/// name of another macro (hereafter called iterator-macro) as an argument followed by a
+/// parenthesized list of extra arguments. The loop-macro expands to a call of the iterator-macro
+/// with a list of comma-separated names of the enum variants wrapped in square brackets, followed
+/// by the extra arguments defined above.
+///
+/// For example, the following code:
+/// ```no_compile
+/// #[derive(ForEachVariant)]
+/// pub enum FooBar {
+///     Foo,
+///     Bar,
+/// }
+/// ```
+/// results in the following macro being defined:
+/// ```
+/// #[macro_export]
+/// macro_rules! for_each_foo_bar_variant {
+///     ( $f:ident($( $args:tt )*) ) => { $f!([Foo, Bar] $($args)*) }
+/// }
+///
+/// pub(crate) use for_each_foo_bar_variant;
+/// ```
+#[proc_macro_derive(ForEachVariant)]
+pub fn derive_for_each_variant(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+    derive_for_each_variant::derive(input)
 }
 
 /// Exposes the function as an application entry point. Entry points are alternative application
